edit user page

+++ app/Http/Controllers/Admin/UsersController.php
@@ -7,27 +7,32 @@ use Inertia\Inertia;
 use App\Http\Requests\Admin\UserRequest;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\UserResource;
+
 class UsersController extends Controller
 {
+    public function __construct()
+    {
+        $this->authorizeResource(User::class);
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
         return Inertia::render(
-            'admin/Users', 
+            'admin/users/Index', 
             [
                 'response' => UserResource::collection(User::with('roles')->paginate())
             ]
         );
     }
 
-    /**
+    /**        
      * Show the form for creating a new resource.
      */
     public function create()
     {
-        //
+        return Inertia::render('admin/users/Create');
     }
 
     /**
@@ -35,7 +40,36 @@ class UsersController extends Controller
      */
     public function store(UserRequest $request)
     {
-        //
+        $validated = $request->validated();
+
+        $data = $request->only([
+            'firstname',
+            'lastname',
+            'name',
+            'email',
+            'password',
+            'role',
+            'avatar',
+            'permissions'
+        ]);
+
+        $user = User::create($data);
+        $validated = $request->validated();
+
+        $data = $request->only([
+            'firstname',
+            'lastname',
+            'name',
+            'email',
+            'password',
+            'role',
+            'avatar',
+            'permissions'
+        ]);
+
+        $user = User::create($data);
+
+        return redirect('admin.user.index');
     }
 
     /**
@@ -51,7 +85,7 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
-        //
+        return Inertia::render('admin/users/Edit', [ 'user' => new UserResource($user) ]);
     }
 
     /**
@@ -59,7 +93,23 @@ class UsersController extends Controller
      */
     public function update(UserRequest $request, User $user)
     {
-        //
+        $validated = $request->validated();
+
+        $data = $request->only([
+            'firstname',
+            'lastname',
+            'name',
+            'email',
+            'password',
+            'role',
+            'avatar',
+            'permissions',
+            'status'
+        ]);
+
+        $user = $user->update($data);
+
+        return to_route('admin.users.index');
     }
 
     /**
@@ -67,6 +117,12 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if ($user->is(auth()->user())) {
+            return $this->errorUnauthorized('You can\'t delete for yourself and other Administrators!');
+        }
+
         $user->delete();
+
+        return $this->respondWithMessage("Successfully deleted");
     }
 }
